Moved Square component and added a separate SquareUtils for utils used in Square component. Issue #6
